+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [9.1.9] - 2020-08-10
+### Bugfix
+- Fixed wrong font size for juwelo icon font
+- update juwelo icon font with new test icon
+
+### Changes
+- extended juwelo icon font component with 2 more props: error, light 
+
 ## [9.1.8] - 2020-08-10
 ### Bugfix
 - Added missing build files